pyproject.toml
[tool.poetry]
name = "cw-aws-infra"
version = "0.1.0"
description = "A Pyhton AWS CDK project to build a AWS Neptune + ECS + Sagemaker stack"
authors = ["Solomon"]
readme = "README.md"
packages = [{include = "cw_aws_infra"}]

[tool.poetry.dependencies]
python = "^3.10"
aws-cdk-lib = "^2.99.1"
constructs = "^10.2.70"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
